Turn guess_spatial_dimensions to a private function  (#4031)
* move  in _gaussian.py

* move related test and remove function from color submodul

* Add warning for deprecation

* Fix unittests

* Fix import

* add stacklevel to warn()
