+++ skimage/filters/__init__.py
@@ -1,5 +1,5 @@
 from .lpi_filter import inverse, wiener, LPIFilter2D
-from ._gaussian import gaussian
+from ._gaussian import gaussian, _guess_spatial_dimensions
 from .edges import (sobel, sobel_h, sobel_v,
                     scharr, scharr_h, scharr_v,
                     prewitt, prewitt_h, prewitt_v,
